phys_to_virt lol merge mistake

+++ src/memory/pmm.zig
@@ -99,7 +99,7 @@ pub fn free_phys(phys: usize, size: usize) void {
   unreachable;
 }
 
-pub fn phys_to_virt(phys: usize) u64 {
+pub fn phys_to_virt(phys: usize) usize {
   return os.memory.paging.kernel_context.phys_to_virt(phys);
 }
